Add the changes needed for the analysis service

+++ Molecules.Core/Domain/ValueObjects/Molecules/AtomOrbital.cs
@@ -6,20 +6,20 @@ namespace Molecules.Core.Domain.ValueObjects.Molecules
     {
         public int Position { get; set; }
         public string Symbol { get; set; } = string.Empty;
-        public decimal? MullikenPopulation { get; set; }
-        public decimal? MullikenPopulationMinus1 { get; set; }
-        public decimal? MullikenPopulationPlus1 { get; set; }
-        public decimal? LowdinPopulation { get; set; }
-        public decimal? LowdinPopulationMinus1 { get; set; }
-        public decimal? LowdinPopulationPlus1 { get; set; }
+        public double? MullikenPopulation { get; set; }
+        public double? MullikenPopulationMinus1 { get; set; }
+        public double? MullikenPopulationPlus1 { get; set; }
+        public double? LowdinPopulation { get; set; }
+        public double? LowdinPopulationMinus1 { get; set; }
+        public double? LowdinPopulationPlus1 { get; set; }
 
         [JsonIgnore]
-        public decimal? MullikenPopulationLumo => MullikenPopulationPlus1 - MullikenPopulation;
+        public double? MullikenPopulationLumo => MullikenPopulationPlus1 - MullikenPopulation;
         [JsonIgnore]
-        public decimal? MullikenPopulationHomo => MullikenPopulation - MullikenPopulationMinus1;
+        public double? MullikenPopulationHomo => MullikenPopulation - MullikenPopulationMinus1;
         [JsonIgnore]
-        public decimal? LowdinPopulationLumo => LowdinPopulationPlus1 - LowdinPopulation;
+        public double? LowdinPopulationLumo => LowdinPopulationPlus1 - LowdinPopulation;
         [JsonIgnore]
-        public decimal? LowdinPopulationHomo => LowdinPopulation - LowdinPopulationMinus1;
+        public double? LowdinPopulationHomo => LowdinPopulation - LowdinPopulationMinus1;
     }
 }
