+++ Molecules.Core/Domain/ValueObjects/Molecules/ElectronicPotential.cs
@@ -3,11 +3,11 @@
     public class ElectronicPotential
     {
         public string? Type { get; set; }
-        public decimal? PosX { get; set; }
-        public decimal? PosY { get; set; }
-        public decimal? PosZ { get; set; }
-        public decimal? Nuclear { get; set; }
-        public decimal? Electronic { get; set; }
-        public decimal? Total { get; set; }
+        public double? PosX { get; set; }
+        public double? PosY { get; set; }
+        public double? PosZ { get; set; }
+        public double? Nuclear { get; set; }
+        public double? Electronic { get; set; }
+        public double? Total { get; set; }
     }
 }
