+++ Molecules.Core/Domain/ValueObjects/Reports/ConfigurationReportItem.cs
@@ -2,7 +2,7 @@
 
 namespace Molecules.Core.Domain.ValueObjects.Reports
 {
-    public record ConfigurationReportItem(string Symbol, decimal? Population, decimal? PopulationFraction)
+    public record ConfigurationReportItem(string Symbol, double? Population, double? PopulationFraction)
     {
         public override string ToString()
         {
