+++ Molecules.Core/Domain/ValueObjects/Reports/GeneralMoleculeReport.cs
@@ -6,17 +6,17 @@
 
         public string AtomID { get; set; } = string.Empty;
 
-        public decimal? CHelpGCharge { get; set; }
+        public double? CHelpGCharge { get; set; }
 
-        public decimal? MullNeutral { get; set; }
+        public double? MullNeutral { get; set; }
 
         public string Configuration { get; set; } = string.Empty;
 
-        public decimal? MullLewisAcid { get; set; }
+        public double? MullLewisAcid { get; set; }
 
         public string ConfigurationLewisAcid { get; set; } = string.Empty;
 
-        public decimal? MullLewisBase { get; set; }
+        public double? MullLewisBase { get; set; }
 
         public string ConfigurationLewisBase { get; set; } = string.Empty;
