+++ Molecules.Core/Domain/ValueObjects/Reports/MoleculeAtomPositionReport.cs
@@ -5,8 +5,8 @@
         public string MoleculeName { get; set; } = "";
         public int AtomPosition { get; set; }
         public string AtomSymbol { get; set; } = "";
-        public decimal? PosX { get; set; }
-        public decimal? PosY { get; set; }
-        public decimal? PosZ { get; set; }
+        public double? PosX { get; set; }
+        public double? PosY { get; set; }
+        public double? PosZ { get; set; }
     }
 }
