+++ Molecules.Core/Domain/ValueObjects/Reports/MoleculeAtomsPopulationReport.cs
@@ -6,16 +6,16 @@
 
         public string AtomID { get; set; } = "";
 
-        public decimal? MullikenPopulation { get; set; }
+        public double? MullikenPopulation { get; set; }
 
-        public decimal? MullikenPopulationHOMO { get; set; }
+        public double? MullikenPopulationHOMO { get; set; }
 
-        public decimal? MullikenPopulationLUMO { get; set; }
+        public double? MullikenPopulationLUMO { get; set; }
 
-        public decimal? LowdinPopulation { get; set; }
+        public double? LowdinPopulation { get; set; }
 
-        public decimal? LowdinPopulationHOMO { get; set; }
+        public double? LowdinPopulationHOMO { get; set; }
 
-        public decimal? LowdinPopulationLUMO { get; set; }
+        public double? LowdinPopulationLUMO { get; set; }
     }
 }
