+++ Molecules.Core/Factories/CalcParsers/FukuiEnergyLewisHOMOParser.cs
@@ -20,7 +20,7 @@
             return EnergyTag;
         }
 
-        internal static decimal GetEnergy(List<string> fileLines)
+        internal static double GetEnergy(List<string> fileLines)
         {
             return new FukuiEnergyLewisHOMOParser().Parse(fileLines);
         }
