+++ Molecules.Shared/XyzConversion.cs
@@ -4,9 +4,9 @@
     {
         private static readonly string[] _returns = ["\r\n", "\r", "\n"];
 
-        public static List<(string symbol, decimal x, decimal y, decimal z)> ParseXyz(string xyz)
+        public static List<(string symbol, double x, double y, double z)> ParseXyz(string xyz)
         {
-            List<(string symbol, decimal x, decimal y, decimal z)> retval = [];
+            List<(string symbol, double x, double y, double z)> retval = [];
 
             string[] lines = xyz.Split(_returns, StringSplitOptions.None);
             foreach (var line in lines)
@@ -15,9 +15,9 @@
                 if (lineItems.Length > 2)
                 {
                     retval.Add((lineItems[0],
-                                    StringConversion.ToDecimal(lineItems[1]),
-                                        StringConversion.ToDecimal(lineItems[2]),
-                                            StringConversion.ToDecimal(lineItems[3])));
+                                    StringConversion.ToDouble(lineItems[1]),
+                                        StringConversion.ToDouble(lineItems[2]),
+                                            StringConversion.ToDouble(lineItems[3])));
                 }
             }
             return retval;
